Huge_Update[CombatSystem]
Combat system upgrade added new child attack core as melee attack and ranged attack

+++ Assets/Scripts/Subsystems/StatsComponent.cs
@@ -5,7 +5,7 @@ public class StatsComponent : MonoBehaviour
     [SerializeField] private BaseStatsSO m_baseStats;
 
     public int MaxHP => m_baseStats.maxHP;
-    public float AttackDamage => m_baseStats.attackDamage;
+    public int AttackDamage => m_baseStats.attackDamage;
     public float AttackRate => m_baseStats.attackRate;
     public float AttackRange => m_baseStats.attackRange;
     public float MoveSpeed => m_baseStats.moveSpeed;
